Move booking related form jsx into a separate component, and move necessary state down
