+++ src/app/book/[url]/page.tsx
@@ -3,8 +3,6 @@
 import { Avatar } from '@/components/UI/Avatar/Avatar';
 import { Card } from '@/components/UI/Card/Card';
 import { Modal } from '@/components/UI/Modal/Modal';
-import { Spinner } from '@/components/UI/Spinner/Spinner';
-import { TextButton } from '@/components/UI/TextButton/TextButton';
 import { NewBusiness } from '@/types/Business';
 import { Service } from '@/types/Service';
 import { Staff } from '@/types/User';
@@ -13,15 +11,12 @@ import { GET_BOOKING_SITE } from '@/utility/queries/bookingSiteQueries';
 import { GET_BUSINESS, GET_BUSINESS_SERVICES, GET_BUSINESS_STAFF } from '@/utility/queries/businessQueries';
 import { useQuery } from '@apollo/client';
 import clsx from 'clsx';
-import { useEffect, useMemo, useState } from 'react';
+import { SetStateAction, useEffect, useMemo, useState } from 'react';
 import styles from './book.module.scss';
-import { Confirm } from './forms/confirm';
-import { SelectAgent } from './forms/selectAgent';
-import { SelectService } from './forms/selectService';
-import { SelectTime } from './forms/selectTime';
 import Loading from './loading';
 import { NotFound } from './notfound';
 import formStyles from './forms/forms.module.scss';
+import { Book } from './book';
 
 export interface SelectedState {
   service: Service | null,
@@ -36,7 +31,6 @@ export interface Details {
 
 export default function Page({params}: { params: any }) {
   const [tab, setTab] = useState('Book');
-  const [formTab, setFormTab] = useState(0);
   const [business, setBusiness] = useState<NewBusiness>();
   const [services, setServices] = useState<Service[]>([]);
   const [staff, setStaff] = useState<Staff[]>([]);
@@ -50,7 +44,6 @@ export default function Page({params}: { params: any }) {
 
   const dateString = useMemo(() => selected.startDate && getDateTimeStringFull(selected.startDate), [selected.startDate]);
 
-  useEffect(() => setFormTab(Object.values(selected).filter(v => !!v).length), [selected]);
 
   // Booking site & business data if available
   const { data: bookingSiteData, loading: loadingBookingSiteData } = useQuery(GET_BOOKING_SITE, { variables: { url: params.url }});
@@ -65,25 +58,6 @@ export default function Page({params}: { params: any }) {
 
   const initialLoading = useMemo(() => loadingBookingSiteData || loadingBusinessData, [loadingBookingSiteData, loadingBusinessData]);
   const businessDataLoading = useMemo(() => loadingServicesData || loadingStaffData, [loadingServicesData, loadingStaffData]);
-
-  const tabs = useMemo(() => {
-    if (!business) return [];
-    
-    const tabs = [];
-    tabs.push(<SelectService key="selectService" services={services} setSelected={setSelected} />)
-
-    if (!selected.service) return tabs;
-    const filteredStaff = staff.filter(s => selected.service!.assigned_staff.find(staff => staff.id === s.id));
-    tabs.push(<SelectAgent key="selectAgent" setSelected={setSelected} businessId={business.id} staff={filteredStaff} />)
-
-    if (!selected.staff) return tabs;
-    tabs.push(<SelectTime key="selectTime" setSelected={setSelected} business={business} selectedStaff={selected.staff} selectedService={selected.service} />);
-
-    if (!selected.startDate) return tabs;
-    tabs.push(<Confirm setSuccessModal={setSuccessDetails} setSelected={setSelected} business={business} selected={selected} key="confirm" />)
-
-    return tabs;
-  }, [business, selected, services, staff]);
 
   if (initialLoading) return <Loading />
   if(!bookingSiteData?.getBookingSite || !business) return <NotFound />
@@ -103,32 +77,7 @@ export default function Page({params}: { params: any }) {
         </div>
         <div className={styles.form}>
           <Card className={styles.formCard}>
-            <div className={styles.formProgress}>
-              {['Select a Service', 'Select an Agent', 'Select a Time', 'Confirm Booking'].map((label, i) => 
-                <div key={label} className={styles.progressLabel}>
-                  <p key={label} className={clsx({[styles.current]: formTab === i})}>{label}</p>
-                  {i !== 3 && <hr />}
-                </div>
-              )}
-            </div>
-            <div className={styles.formShowing}>
-              {businessDataLoading ? <Spinner style={{margin: 30}} /> : tabs[formTab]}
-            </div>
-            <div className={clsx(styles.formNavigate, 'noselect')}>
-              {!formTab || businessDataLoading ? <div /> : 
-                <TextButton className={styles.back}
-                  onClick={() => {
-                    setSelected(p => ({
-                      ...p,
-                      [Object.keys(p)[formTab - 1]]: null,
-                    }))
-                    setFormTab(p => p - 1);
-                  }}
-                >
-                  Go Back
-                </TextButton>
-              }
-            </div>
+            <Book selected={selected} business={business} loadingData={businessDataLoading} setSelected={setSelected} staff={staff} services={services} setSuccessDetails={setSuccessDetails}  />
           </Card>
         </div>
       </div>
